Simplify dictation pipeline and settings autosave
- share transcript cleanup/paste handling between normal and streaming
  dictation
- prevent duplicate local model load/warmup races
- remove unused pill state mode/origin bookkeeping
- make detached music pause work explicit
- clean up small clippy warnings
- flush pending settings autosaves on close/unmount

+++ src-tauri/src/platform/macos/audio_devices.rs
@@ -78,7 +78,7 @@ impl AudioPropertyListener {
             )
         };
 
-        if status != kAudioHardwareNoError as i32 {
+        if status != kAudioHardwareNoError {
             return Err(format!("CoreAudio listener failed with status {status}"));
         }
